Static analysis cleanups, static libmagic
