+++ include/osquery/core.h
@@ -240,10 +240,9 @@ inline size_t incUtf8StringIterator(_Iterator1& it, const _Iterator2& last) {
     return 0;
   }
 
-  unsigned char c;
   size_t res = 1;
   for (++it; last != it; ++it, ++res) {
-    c = *it;
+    unsigned char c = *it;
     if (!(c & 0x80) || ((c & 0xC0) == 0xC0)) {
       break;
     }
