+++ osquery/config/tests/config_tests.cpp
@@ -134,26 +134,6 @@ class TestConfigParserPlugin : public ConfigParserPlugin {
 // An intermediate boolean to check parser updates.
 bool TestConfigParserPlugin::update_called = false;
 
-TEST_F(ConfigTests, test_splay) {
-  auto val1 = splayValue(100, 10);
-  EXPECT_GE(val1, 90);
-  EXPECT_LE(val1, 110);
-
-  auto val2 = splayValue(100, 10);
-  EXPECT_GE(val2, 90);
-  EXPECT_LE(val2, 110);
-
-  auto val3 = splayValue(10, 0);
-  EXPECT_EQ(val3, 10);
-
-  auto val4 = splayValue(100, 1);
-  EXPECT_GE(val4, 99);
-  EXPECT_LE(val4, 101);
-
-  auto val5 = splayValue(1, 10);
-  EXPECT_EQ(val5, 1);
-}
-
 TEST_F(ConfigTests, test_parse) {
   auto c = Config();
   auto tree = getExamplePacksConfig();
